move reverse_int() into reverse_int.c, and generate init.c via tools::package_native_routine_registration_skeleton(".",  "src/init.c")
