+++ NAMESPACE
@@ -1,10 +1,13 @@
+# Generated by roxygen2: do not edit by hand
+
 S3method("+",Foo)
 S3method(hello,character)
 S3method(hello,default)
 export(hello)
-export(reverse, reverse_int)
+export(reverse)
+export(reverse_int)
 export(rock_fun)
 export(split_filename)
 importFrom(tools,file_ext)
 importFrom(tools,file_path_sans_ext)
-useDynLib(rmini, .registration = TRUE, .fixes = "rmini_")
+useDynLib(rmini, .registration = TRUE)
